KeyManager: update to vita-ske-1h
A new version of Vita’s simple key exchange protocol and its implementation.

Changes:

Instead of using parts of the derived ephemeral key material as salts (as used
in RFC 4106) derive the salts from the authenticated nonces.Rationale: the
salts are public and it was suspicious and confusing that they were part of the
negotiated secret keys when they need not to be. Instead we do the obvious and
use part of the authenticated, unpredictable nonces that are exchanged during
the protocol for salts, and derive four bytes less ephemeral key material for
each key.