KeyManager: add delay before activating new outbound SAs...
...to give the other node time to install its matching RX SA. Like

  10264f68 KeyManager: add negotiation_delay and protocol jitter

this is a hack that needs proper resolving through design changes in the
protocol fsm. It intends to solve the issue of transmitting packets on a SA the
receiving node has not yet had time to install, and partially solve another
issue where one of the two nodes that partake in a key exchange fails to
complete the exchange (i.e., because it does not receive the final protocol
message due to packet loss) in which case (for a limited duration) the other
node might transmit packets on a SA that is not active on that node until a new
SA is negotiated.

Again, a future version of the AKE protocol should be designed with these
failure modes in mind.