vita-ske2a: design and implementation
Bumps the vita-ake protocol major version...

The protocol fsm is split up to model a responder and an initiator, where the
initiator begins by requesting a nonced challenge from the responder. Only an
authentic challenges will be considered by the initiator, and enable it to
craft an authentic nonce_key message to be accepted by the responder.

See fsm-protocol.png

These changes address issues with the previous protocol fsm which required a
lot of quirky logic and jitter hacks to ensure it wouldn’t clog up (and there
was a remaining defect where spoofing an initial nonce message in between two
idle peers would cause them to DoS each other, OOPS.)

Cryptographically, the new protocol version should be equivalent, with the
exception that participants can no longer be tricked into accepting nonces from
unauthenticated peers.

On the KeyManager side, SAs are now maintained in a set of queues in an effort
to prepare it for scaling Vita by negotiating multiple SA pairs for each route.
Generally speaking, the new protocol allowed to simplify and streamline a few
tricky parts of the implementation. Notably missing, as of yet, is a strategy
to deal with compounded packet loss in a multiple SA per route setup.