vita/exchange: include initiator/responder addresses into prologue
This addresses an issue in the AKE protocol where a party could be tricked to
authenticate with itself, i.e.

> A subtle attack of PSK based authentication is the possibility of a party
> authenticating with itself without its awareness, as described in Selfie:
> reflections on TLS 1.3 with PSK.

By including the network identities of both initiator and responder in the
Noise prologue, we ensure that both parties agree on the the intent of the
exchange. E.g., in a testing scenario a node can use its address in both the
initiator and responder fields and thereby encode the intent to negotiatiate
with itself, while in a production scenario the one of the fields would contain
the address of the intended peer so that an exchange with a different peer
fails unless it includes matching network identities in its prologue.

Note that the network address itself is not treated as a trusted identity.
It is merely used as a signal of intent by a party that authenticates itself
using the correct PSK.

This change also adresses an issue where the behaviour of non-symmetric work
queue configurations was ill-defined. With this change AKE negotiations will
fail unless both peers have configured each other as gateways for a given route
explicitly.

This change bumps the VITA-SKE protocol version to 5.