vita: implement worker monitoring with manager/core.worker
This is not as reliable as the previous strategy (a worker could die, then be
be reaped due to a configuration change, and not cause a manager exit), but
should be good enough and more selective wrt to miscellaneous subprocesses.

Good enough because it maintains the basic promise that the manager will exit
if it ever finds itself with less workers than it thinks it should have.